+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/services/AbsoluteUrlProvider.kt
@@ -3,7 +3,7 @@ package uk.gov.communities.prsdb.webapp.services
 import org.springframework.stereotype.Service
 import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder
 import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder.on
-import uk.gov.communities.prsdb.webapp.controllers.LandlordDashboardController
+import uk.gov.communities.prsdb.webapp.controllers.LandlordController
 import uk.gov.communities.prsdb.webapp.controllers.RegisterLAUserController
 import java.net.URI
 
@@ -11,7 +11,7 @@ import java.net.URI
 class AbsoluteUrlProvider {
     fun buildLandlordDashboardUri(): URI =
         MvcUriComponentsBuilder
-            .fromMethodCall(on(LandlordDashboardController::class.java).index())
+            .fromMethodCall(on(LandlordController::class.java).index())
             .build()
             .toUri()
